phase5/requirements_ml.txt
# Phase 5 ML Requirements - Lightweight scikit-learn-based

# Core ML
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.3

# Data Processing
scipy==1.11.4

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# Model Persistence
joblib==1.3.2

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Utilities
python-dateutil==2.8.2
